# packages/contracts/openapi/social-platform.v1.yaml
openapi: 3.1.0
info:
  title: Swypik Social Platform API
  version: 1.0.0
  description: Modular monorepo platform API contract for feed, events, video, social, cart, checkout, payments, and moderation.
servers:
  - url: http://localhost:8080
    description: Local platform-api
...
paths:
...
  /v1/feed:
    get:
      tags: [feed]
      operationId: listFeed
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: category_id
          in: query
          schema:
            type: string
        - name: creator_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Ranked feed page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedResponse"
...
components:
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 24
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
...
  schemas:
...
    FeedResponse:
      type: object
      required: [items, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FeedItem"
        limit:
          type: integer
        offset:
          type: integer
    FeedItem:
      type: object
      required: [id, has_video, rank_score]
      properties:
        id:
          type: string
        product_id:
          type: string
        creator_id:
          type: string
        category_id:
          type: string
        title:
          type: string
        description:
          type: string
        has_video:
          type: boolean
        video_url:
          type: string
          format: uri
        poster_url:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
        price_ron:
          type: integer
        old_price_ron:
          type: integer
        rating:
          type: number
        orders_count:
          type: integer
        likes_count:
          type: integer
        views_count:
          type: integer
        comments_count:
          type: integer
        published_at:
          type: string
          format: date-time
        rank_score:
          type: number
...